refactor file and folder names
