+++ packages/refine/src/components/PageShow/index.ts
@@ -1,0 +1,1 @@
+export * from './PageShow';
