+++ packages/refine/src/components/ShowContent/index.ts
@@ -1,0 +1,2 @@
+export * from './fields';
+export * from './ShowContent';
