+++ packages/refine/src/components/WorkloadPodsTable/index.ts
@@ -1,0 +1,1 @@
+export * from './WorkloadPodsTable';
