+++ packages/refine/src/model/index.ts
@@ -1,0 +1,2 @@
+export * from './pod-model';
+export * from './resource-model';
