+++ packages/refine/src/pages/configmap/list/index.tsx
@@ -5,15 +5,15 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { CreateButton } from 'src/components/CreateButton';
 import { DeleteManyButton } from 'src/components/DeleteManyButton';
-import { DrawerShow } from 'src/components/Show/DrawerShow';
+import { DrawerShow } from 'src/components/DrawerShow';
 import Table, { IDObject } from 'src/components/Table';
 import { useEagleTable } from 'src/hooks/useEagleTable';
 import {
   NameColumnRenderer,
   NameSpaceColumnRenderer,
   AgeColumnRenderer,
-} from 'src/hooks/useEagleTable/Columns';
-import { Tags } from '../../../components/Show/Tags';
+} from 'src/hooks/useEagleTable/columns';
+import { Tags } from '../../../components/Tags';
 
 export const ConfigmapList: React.FC<IResourceComponentsProps> = <
   T extends IDObject & Unstructured,
